migrate Image Tasks
