+++ src/Mobile.BuildTools.Images/Drawing/FloatExtensions.cs
