+++ src/Mobile.BuildTools.Images/Drawing/VectorImage.cs
