+++ src/Mobile.BuildTools.Images/Drawing/Watermark.cs
