+++ src/Mobile.BuildTools.Images/Drawing/WatermarkImage.cs
