+++ src/Mobile.BuildTools.Images/Drawing/WatermarkTextBanner.cs
