+++ src/Mobile.BuildTools.Images/Tasks/CollectImageAssetsTask.cs
