+++ src/Mobile.BuildTools/Mobile.BuildTools.csproj
@@ -42,11 +42,11 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Mobile.BuildTools.AppSettings\Mobile.BuildTools.AppSettings.csproj" 
+    <ProjectReference Include="..\Mobile.BuildTools.AppSettings\Mobile.BuildTools.AppSettings.csproj"
                       ReferenceOutputAssembly="false" />
-    <ProjectReference Include="..\Mobile.BuildTools.FormsCss\Mobile.BuildTools.FormsCss.csproj" 
+    <ProjectReference Include="..\Mobile.BuildTools.FormsCss\Mobile.BuildTools.FormsCss.csproj"
                       ReferenceOutputAssembly="false" />
-    <ProjectReference Include="..\Mobile.BuildTools.Images\Mobile.BuildTools.Images.csproj" 
+    <ProjectReference Include="..\Mobile.BuildTools.Images\Mobile.BuildTools.Images.csproj"
                       ReferenceOutputAssembly="false" />
     <ProjectReference Include="..\Mobile.BuildTools.Reference\Mobile.BuildTools.Reference.csproj" PackFolder="build" />
     <ProjectReference Include="..\Mobile.BuildTools.XamarinSdk\Mobile.BuildTools.XamarinSdk.csproj" PackFolder="build" />
